refactor(runtime): extract app runtime event handlers
