+++ crates/gwt/src/app_runtime/mod.rs
@@ -145,6 +145,7 @@ mod board;
 mod migration;
 pub(crate) mod persist_dispatcher;
 mod profile;
+mod runtime_events;
 mod title_sync;
 mod ui_trace;
 mod window;
@@ -4683,204 +4684,6 @@ impl AppRuntime {
             .collect()
     }
 
-    pub(crate) fn handle_runtime_output(
-        &mut self,
-        id: String,
-        data: Vec<u8>,
-    ) -> Vec<OutboundEvent> {
-        self.handle_runtime_output_inner(id, data, true)
-    }
-
-    pub(crate) fn handle_daemon_runtime_output(
-        &mut self,
-        id: String,
-        data: Vec<u8>,
-    ) -> Vec<OutboundEvent> {
-        self.handle_runtime_output_inner(id, data, false)
-    }
-
-    fn handle_runtime_output_inner(
-        &mut self,
-        id: String,
-        data: Vec<u8>,
-        publish_to_daemon: bool,
-    ) -> Vec<OutboundEvent> {
-        let Some(address) = self.window_lookup.get(&id).cloned() else {
-            return Vec::new();
-        };
-        if publish_to_daemon {
-            if let Some(tab) = self.tab(&address.tab_id) {
-                publish_runtime_output_change(&tab.project_root, &id, &data);
-            }
-        }
-        vec![OutboundEvent::broadcast(BackendEvent::TerminalOutput {
-            id,
-            data_base64: base64::engine::general_purpose::STANDARD.encode(data),
-        })]
-    }
-
-    pub(crate) fn handle_runtime_status(
-        &mut self,
-        id: String,
-        status: WindowProcessStatus,
-        detail: Option<String>,
-    ) -> Vec<OutboundEvent> {
-        self.handle_runtime_status_inner(id, status, detail, true)
-    }
-
-    pub(crate) fn handle_daemon_runtime_status(
-        &mut self,
-        id: String,
-        status: WindowProcessStatus,
-        detail: Option<String>,
-    ) -> Vec<OutboundEvent> {
-        self.handle_runtime_status_inner(id, status, detail, false)
-    }
-
-    fn handle_runtime_status_inner(
-        &mut self,
-        id: String,
-        status: WindowProcessStatus,
-        detail: Option<String>,
-        publish_to_daemon: bool,
-    ) -> Vec<OutboundEvent> {
-        let Some(_address) = self.window_lookup.get(&id).cloned() else {
-            self.remove_window_state_tracking(&id);
-            self.deregister_pty_writer(&id);
-            self.runtimes.remove(&id);
-            self.window_details.remove(&id);
-            return Vec::new();
-        };
-        if publish_to_daemon {
-            if let Some(address) = self.window_lookup.get(&id) {
-                if let Some(tab) = self.tab(&address.tab_id) {
-                    publish_runtime_status_change(&tab.project_root, &id, status, detail.clone());
-                }
-            }
-        }
-
-        self.window_pty_statuses.insert(id.clone(), status);
-        let composed_status = self.recompute_window_state(&id).unwrap_or(status);
-        let should_auto_close =
-            should_auto_close_agent_window(&self.active_agent_sessions, &id, &composed_status);
-        match detail.as_ref() {
-            Some(detail) if !detail.is_empty() => {
-                self.window_details.insert(id.clone(), detail.clone());
-            }
-            _ => {
-                self.window_details.remove(&id);
-            }
-        }
-        if should_auto_close {
-            self.stop_window_runtime(&id);
-            self.remove_window_state_tracking(&id);
-            if !close_window_from_workspace(
-                &mut self.tabs,
-                &mut self.window_lookup,
-                &mut self.window_details,
-                &id,
-            ) {
-                return Vec::new();
-            }
-            let _ = self.persist();
-            let mut events = Vec::new();
-            self.push_workspace_and_active_work_projection_broadcasts(&mut events);
-            return events;
-        }
-        if matches!(
-            status,
-            WindowProcessStatus::Error | WindowProcessStatus::Stopped
-        ) {
-            self.runtimes.remove(&id);
-            self.remove_window_state_tracking(&id);
-            self.mark_agent_session_stopped(&id);
-        }
-        let _ = self.persist();
-
-        let mut events = Vec::new();
-        if matches!(
-            status,
-            WindowProcessStatus::Error | WindowProcessStatus::Stopped
-        ) {
-            if let Some(event) = self.active_work_projection_broadcast_for_active_tab() {
-                events.push(event);
-            }
-        }
-        events.extend(Self::status_events(id, composed_status, detail));
-        events
-    }
-
-    pub(crate) fn handle_runtime_hook_event(
-        &mut self,
-        event: gwt::RuntimeHookEvent,
-    ) -> Vec<OutboundEvent> {
-        self.handle_runtime_hook_event_inner(event, true)
-    }
-
-    pub(crate) fn handle_daemon_runtime_hook_event(
-        &mut self,
-        event: gwt::RuntimeHookEvent,
-    ) -> Vec<OutboundEvent> {
-        self.handle_runtime_hook_event_inner(event, false)
-    }
-
-    fn handle_runtime_hook_event_inner(
-        &mut self,
-        event: gwt::RuntimeHookEvent,
-        publish_to_daemon: bool,
-    ) -> Vec<OutboundEvent> {
-        if publish_to_daemon {
-            if let Some(project_root) = event.project_root.as_deref().map(PathBuf::from) {
-                publish_runtime_hook_change(&project_root, &event);
-            }
-        }
-        let mut events = vec![OutboundEvent::broadcast(BackendEvent::RuntimeHookEvent {
-            event: event.clone(),
-        })];
-        let Some(window_id) = self.active_window_for_runtime_event(&event) else {
-            return events;
-        };
-        let Some(hook_state) = gwt::window_state::runtime_hook_window_state(&event) else {
-            return events;
-        };
-        self.window_hook_states
-            .insert(window_id.clone(), hook_state);
-        let Some(composed_state) = self.recompute_window_state(&window_id) else {
-            return events;
-        };
-        let should_auto_close = should_auto_close_agent_window(
-            &self.active_agent_sessions,
-            &window_id,
-            &composed_state,
-        );
-        let detail = self.window_details.get(&window_id).cloned();
-        if should_auto_close {
-            self.stop_window_runtime(&window_id);
-            self.remove_window_state_tracking(&window_id);
-            if close_window_from_workspace(
-                &mut self.tabs,
-                &mut self.window_lookup,
-                &mut self.window_details,
-                &window_id,
-            ) {
-                let _ = self.persist();
-                self.push_workspace_and_active_work_projection_broadcasts(&mut events);
-            }
-            return events;
-        }
-        let _ = self.persist();
-        if matches!(
-            composed_state,
-            WindowProcessStatus::Error | WindowProcessStatus::Stopped
-        ) {
-            if let Some(event) = self.active_work_projection_broadcast_for_active_tab() {
-                events.push(event);
-            }
-        }
-        events.extend(Self::status_events(window_id, composed_state, detail));
-        events
-    }
-
     pub(crate) fn handle_launch_complete(
         &mut self,
         window_id: String,
@@ -6265,209 +6068,6 @@ fn update_issue_branch_link_with_cache_dir(
         .map_err(|error| format!("failed to write issue linkage store: {error}"))
 }
 
-#[cfg(unix)]
-const RUNTIME_DAEMON_PUBLISH_QUEUE_CAPACITY: usize = 4096;
-
-#[cfg(unix)]
-enum RuntimeDaemonPublish {
-    Output {
-        project_root: PathBuf,
-        id: String,
-        data: Vec<u8>,
-    },
-    Status {
-        project_root: PathBuf,
-        id: String,
-        status: WindowProcessStatus,
-        detail: Option<String>,
-    },
-    Hook {
-        project_root: PathBuf,
-        event: gwt::RuntimeHookEvent,
-    },
-}
-
-#[cfg(unix)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RuntimeDaemonPublishEnqueueError {
-    Full,
-    Disconnected,
-}
-
-#[cfg(unix)]
-static RUNTIME_DAEMON_PUBLISH_QUEUE: std::sync::OnceLock<
-    Option<std_mpsc::SyncSender<RuntimeDaemonPublish>>,
-> = std::sync::OnceLock::new();
-
-#[cfg(unix)]
-fn runtime_daemon_publish_sender() -> Option<&'static std_mpsc::SyncSender<RuntimeDaemonPublish>> {
-    RUNTIME_DAEMON_PUBLISH_QUEUE
-        .get_or_init(|| {
-            let (sender, receiver) = std_mpsc::sync_channel(RUNTIME_DAEMON_PUBLISH_QUEUE_CAPACITY);
-            match std::thread::Builder::new()
-                .name("gwt-runtime-daemon-publish-worker".to_string())
-                .spawn(move || run_runtime_daemon_publish_worker(receiver))
-            {
-                Ok(_handle) => Some(sender),
-                Err(err) => {
-                    tracing::debug!(error = %err, "runtime daemon publish worker spawn failed");
-                    None
-                }
-            }
-        })
-        .as_ref()
-}
-
-#[cfg(unix)]
-fn run_runtime_daemon_publish_worker(receiver: std_mpsc::Receiver<RuntimeDaemonPublish>) {
-    for publish in receiver {
-        publish_runtime_daemon_event(publish);
-    }
-}
-
-#[cfg(unix)]
-fn try_enqueue_runtime_daemon_publish(
-    sender: &std_mpsc::SyncSender<RuntimeDaemonPublish>,
-    publish: RuntimeDaemonPublish,
-) -> Result<(), RuntimeDaemonPublishEnqueueError> {
-    sender.try_send(publish).map_err(|err| match err {
-        std_mpsc::TrySendError::Full(_) => RuntimeDaemonPublishEnqueueError::Full,
-        std_mpsc::TrySendError::Disconnected(_) => RuntimeDaemonPublishEnqueueError::Disconnected,
-    })
-}
-
-#[cfg(unix)]
-fn enqueue_runtime_daemon_publish(publish: RuntimeDaemonPublish) {
-    let Some(sender) = runtime_daemon_publish_sender() else {
-        return;
-    };
-    if let Err(err) = try_enqueue_runtime_daemon_publish(sender, publish) {
-        tracing::debug!(
-            ?err,
-            "runtime daemon publish queue rejected event (non-fatal)"
-        );
-    }
-}
-
-#[cfg(unix)]
-fn publish_runtime_daemon_event(publish: RuntimeDaemonPublish) {
-    match publish {
-        RuntimeDaemonPublish::Output {
-            project_root,
-            id,
-            data,
-        } => {
-            let payload =
-                gwt::runtime_daemon_events::runtime_output_payload(&id, &data, std::process::id());
-            let result = gwt::daemon_publisher::publish_event(
-                &project_root,
-                gwt::runtime_daemon_events::RUNTIME_OUTPUT_CHANNEL,
-                payload,
-            );
-            if let Err(err) = result {
-                tracing::debug!(
-                    error = %err,
-                    project_root = %project_root.display(),
-                    window_id = %id,
-                    "runtime output daemon publish failed (non-fatal)"
-                );
-            }
-        }
-        RuntimeDaemonPublish::Status {
-            project_root,
-            id,
-            status,
-            detail,
-        } => {
-            let payload = gwt::runtime_daemon_events::runtime_status_payload(
-                &id,
-                status,
-                detail,
-                std::process::id(),
-            );
-            let result = gwt::daemon_publisher::publish_event(
-                &project_root,
-                gwt::runtime_daemon_events::RUNTIME_STATUS_CHANNEL,
-                payload,
-            );
-            if let Err(err) = result {
-                tracing::debug!(
-                    error = %err,
-                    project_root = %project_root.display(),
-                    window_id = %id,
-                    "runtime status daemon publish failed (non-fatal)"
-                );
-            }
-        }
-        RuntimeDaemonPublish::Hook {
-            project_root,
-            event,
-        } => {
-            let payload =
-                gwt::runtime_daemon_events::runtime_hook_payload(&event, std::process::id());
-            let result = gwt::daemon_publisher::publish_event(
-                &project_root,
-                gwt::runtime_daemon_events::RUNTIME_HOOK_CHANNEL,
-                payload,
-            );
-            if let Err(err) = result {
-                tracing::debug!(
-                    error = %err,
-                    project_root = %project_root.display(),
-                    "runtime hook daemon publish failed (non-fatal)"
-                );
-            }
-        }
-    }
-}
-
-#[cfg(unix)]
-fn publish_runtime_output_change(project_root: &Path, id: &str, data: &[u8]) {
-    enqueue_runtime_daemon_publish(RuntimeDaemonPublish::Output {
-        project_root: project_root.to_path_buf(),
-        id: id.to_string(),
-        data: data.to_vec(),
-    });
-}
-
-#[cfg(not(unix))]
-fn publish_runtime_output_change(_project_root: &Path, _id: &str, _data: &[u8]) {}
-
-#[cfg(unix)]
-fn publish_runtime_status_change(
-    project_root: &Path,
-    id: &str,
-    status: WindowProcessStatus,
-    detail: Option<String>,
-) {
-    enqueue_runtime_daemon_publish(RuntimeDaemonPublish::Status {
-        project_root: project_root.to_path_buf(),
-        id: id.to_string(),
-        status,
-        detail,
-    });
-}
-
-#[cfg(not(unix))]
-fn publish_runtime_status_change(
-    _project_root: &Path,
-    _id: &str,
-    _status: WindowProcessStatus,
-    _detail: Option<String>,
-) {
-}
-
-#[cfg(unix)]
-fn publish_runtime_hook_change(project_root: &Path, event: &gwt::RuntimeHookEvent) {
-    enqueue_runtime_daemon_publish(RuntimeDaemonPublish::Hook {
-        project_root: project_root.to_path_buf(),
-        event: event.clone(),
-    });
-}
-
-#[cfg(not(unix))]
-fn publish_runtime_hook_change(_project_root: &Path, _event: &gwt::RuntimeHookEvent) {}
-
 fn maybe_register_codex_managed_hook_trust_for_launch(
     profile_config_path: &Path,
     worktree_path: &Path,
@@ -6630,35 +6230,6 @@ mod tests {
         ProjectTabRuntime, UserEvent, WindowRuntime, WorkspaceResumeContext,
     };
     use crate::{combined_window_id, geometry_to_pty_size, same_worktree_path, PtyWriterRegistry};
-
-    #[cfg(unix)]
-    #[test]
-    fn runtime_daemon_publish_enqueue_is_bounded_and_nonblocking() {
-        let (sender, _receiver) = mpsc::sync_channel(1);
-        let project_root = PathBuf::from("/tmp/gwt-project");
-
-        assert!(super::try_enqueue_runtime_daemon_publish(
-            &sender,
-            super::RuntimeDaemonPublish::Output {
-                project_root: project_root.clone(),
-                id: "tab-1::shell-1".to_string(),
-                data: b"first".to_vec(),
-            },
-        )
-        .is_ok());
-        assert!(matches!(
-            super::try_enqueue_runtime_daemon_publish(
-                &sender,
-                super::RuntimeDaemonPublish::Status {
-                    project_root,
-                    id: "tab-1::shell-1".to_string(),
-                    status: WindowProcessStatus::Running,
-                    detail: None,
-                },
-            ),
-            Err(super::RuntimeDaemonPublishEnqueueError::Full)
-        ));
-    }
 
     #[derive(Debug, Clone)]
     struct CapturedTracingEvent {
